chore(bluetooth-transport): platform specific code as trait
